Extract the priority queue module
