+++ packages/priority-queue/CHANGELOG.md
@@ -1,0 +1,3 @@
+### 1.0.0 (Unreleased)
+
+Initial release.
