+++ packages/priority-queue/package.json
@@ -1,0 +1,29 @@
+{
+	"name": "@wordpress/priority-queue",
+	"version": "1.0.0-alpha.0",
+	"description": "Generic browser priority queue.",
+	"author": "The WordPress Contributors",
+	"license": "GPL-2.0-or-later",
+	"keywords": [
+		"wordpress",
+		"browser",
+		"async"
+	],
+	"homepage": "https://github.com/WordPress/gutenberg/tree/master/packages/priority-queue/README.md",
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/WordPress/gutenberg.git"
+	},
+	"bugs": {
+		"url": "https://github.com/WordPress/gutenberg/issues"
+	},
+	"main": "build/index.js",
+	"module": "build-module/index.js",
+	"react-native": "src/index",
+	"dependencies": {
+		"@babel/runtime": "^7.0.0"
+	},
+	"publishConfig": {
+		"access": "public"
+	}
+}
